dreamblog/requirements.txt
Django==2.2.28
django-crispy-forms==1.14.0
django-tinymce==3.2.0
Pillow==9.5.0
